requirements.txt
torch
torchvision
pandas
numpy
matplotlib
seaborn
scikit-learn
streamlit
pillow